Release 08/05

+++ src/features/asset/requests.ts
@@ -15,6 +15,8 @@ import { assetActionsCreators } from './slice';
 import { ASSET_STORAGE_BUCKET } from '@/constants/asset';
 
 export async function requestDeleteFiles(data: RequestDeleteFilesReq): Promise<any> {
+    if (!data.deleteKeys.length) return;
+
     const res = await apiService.delete('/assets/request/deleteFile', data);
     return res;
 }
@@ -72,7 +74,7 @@ export async function getAsset(): Promise<GetAssetApiRes> {
 }
 
 export async function sendRequestUpload(data: AssetSendRequestUploadReq): Promise<AssetSendRequestUploadApiRes> {
-    const res = apiService.post<string>('/assets/request/upload', data);
+    const res = await apiService.post<string>('/assets/request/upload', data);
     return res;
 }
 
@@ -84,4 +86,9 @@ export async function signingMediaC2PA(data: SigningMediaC2PAReq): Promise<Axios
         creator: data.creator,
         filename: data.filename,
     });
+}
+
+export async function extractAssetColors(id: string) {
+    const res = await apiService.get<string[]>(`/assets/${id}/colors`);
+    return res;
 }
